Fix TreeGraph breaking on blank commit messages
The regex for parsing commits didn't match if a commit had a blank message. That resulted in the entire commit line being passed to gitgraph.js instead of just the graph representation, and gitgraph then incorrectly renders a giant canvas that breaks the layout.

Fix is simply to make the commit message optional in the regex.